requirements.txt
anthropic==0.49.0
beautifulsoup4==4.13.3
langchain==0.3.21
langchain_chroma==0.2.2
langchain_community==0.3.20
langchain_core==0.3.45
langchain_openai==0.3.9
langchain_text_splitters==0.3.7
matplotlib==3.9.4
nltk==3.9.1
numpy==1.26.2
openai==1.66.5
pandas==2.2.3
pypdf==5.4.0
python-dotenv==1.0.1
scikit_learn==1.6.1
sentence_transformers==3.3.1
spacy==3.8.3
st_annotated_text==4.0.2
streamlit==1.41.1
streamlit_pdf_viewer==0.0.21
tiktoken==0.8.0
torch==2.5.1+cu124
transformers==4.44.2
